# src/shop/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация для API интернет-магазина",
        "title": "Shop API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/supplier": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "создать поставщика",
                "parameters": [
                    {
                        "description": "Создаваемый поставщик",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SupplierCreateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SupplierResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
        "/supplier/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "получить поставщика по id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SupplierResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "404": {
                        "description": "Client not found",
                        "schema": {
                            "$ref": "#/definitions/dto.Error404"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "удалить поставщика по id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request: invalid JSON or validation failed",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "404": {
                        "description": "client not found",
                        "schema": {
                            "$ref": "#/definitions/dto.Error404"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "обновить адрес поставщика по id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновляемые поля",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SupplierUpdateAddressRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SupplierResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request: invalid JSON or validation failed",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "404": {
                        "description": "client not found",
                        "schema": {
                            "$ref": "#/definitions/dto.Error404"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
        "/suppliers": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "получить всех поставщиков",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuppliersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request: invalid JSON or validation failed",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SuppliersNotFound"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddressCreateDTO": {
            "type": "object",
            "required": [
                "city",
                "country",
                "street"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "example": "London"
                },
                "country": {
                    "type": "string",
                    "example": "UK"
                },
                "street": {
                    "type": "string",
                    "example": "Privet Drive"
                }
            }
        },
        "dto.AddressDTO": {
            "type": "object",
            "required": [
                "city",
                "country",
                "street"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "example": "London"
                },
                "country": {
                    "type": "string",
                    "example": "UK"
                },
                "id": {
                    "type": "string",
                    "example": "a123b456-c789-d012-e345-67890abcdef1"
                },
                "street": {
                    "type": "string",
                    "example": "Privet Drive"
                }
            }
        },
...
        "dto.Error400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "status": {
                    "type": "string",
                    "example": "invalid JSON"
                }
            }
        },
        "dto.Error404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "status": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
        "dto.Error500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "status": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
        "dto.SupplierCreateRequestDTO": {
            "type": "object",
            "required": [
                "name",
                "phone"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/dto.AddressCreateDTO"
                },
                "name": {
                    "type": "string",
                    "example": "Magic Supplies Inc."
                },
                "phone": {
                    "type": "string",
                    "example": "+44-123-456-789"
                }
            }
        },
        "dto.SupplierResponseDTO": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/dto.AddressDTO"
                },
                "address_id": {
                    "type": "string",
                    "example": "address-5678"
                },
                "id": {
                    "type": "string",
                    "example": "supplier-1234"
                },
                "name": {
                    "type": "string",
                    "example": "Magic Supplies Inc."
                },
                "phone": {
                    "type": "string",
                    "example": "+44-123-456-789"
                }
            }
        },
        "dto.SupplierUpdateAddressRequestDTO": {
            "type": "object",
            "required": [
                "city",
                "country",
                "street"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Edinburgh"
                },
                "country": {
                    "type": "string",
                    "example": "UK"
                },
                "street": {
                    "type": "string",
                    "example": "Royal Mile"
                }
            }
        },
        "dto.SuppliersNotFound": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "suppliers": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                }
            }
        },
        "dto.SuppliersResponse": {
            "type": "object",
            "properties": {
                "suppliers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SupplierResponseDTO"
                    }
                }
            }
        }
    }
}